+++ backend/canisters/user/impl/src/openchat_bot.rs
@@ -74,5 +74,5 @@ pub(crate) fn send_message_with_reply(
         now: state.env.now(),
     };
 
-    handle_message_impl(args, None, false, state)
+    handle_message_impl(args, state)
 }